Improvements in PatternGenerator, second version of refHit selection
PatternGenerator:
implemented option to do the histograms of deltaPhiVsPhiRef,
fixed bug in PatternGenerator::updateStatUsingMatcher2()
(refLayerLogicNumber -> refLayer)

PatternOptimizerBase::savePatternsInRoot - modification for drawing
deltaPhiVsPhiRef histograms

OMTFinput::getRefHitStubs - alternative version of refHit selection
logic,
related changes in OMTFProcessor<GoldenPatternType>::processInput and
OMTFinput.

+++ L1Trigger/L1TMuonOverlapPhase1/src/ProcConfigurationBase.cc
@@ -77,7 +77,7 @@ void ProcConfigurationBase::configureFromEdmParameterSet(const edm::ParameterSet
   }
 
   if (edmParameterSet.exists("dtPhiBUnitsRad")) {
-    dtPhiBUnitsRad_ = edmParameterSet.getParameter<double>("dtPhiBUnitsRad");
+    dtPhiBUnitsRad_ = edmParameterSet.getParameter<int>("dtPhiBUnitsRad");
     edm::LogVerbatim("OMTFReconstruction") << "dtPhiBUnitsRad: " << dtPhiBUnitsRad_ << std::endl;
   }
 }
